refactor(transport): streamline transport implementation
* Remove RestTransport and related REST logic.
* Update transport to always use StreamTransport.
* Refactor API calls to use new models API.
* Enhance type definitions for better clarity.
